+++ Integration/Orleans.InProcess/for_Reactors/when_handling_event/and_it_fails/but_not_second_time.cs
@@ -5,14 +5,15 @@ using Cratis.Chronicle.Contracts.Jobs;
 using Cratis.Chronicle.Contracts.Observation;
 using Cratis.Chronicle.Events;
 using Cratis.Chronicle.Integration.Base;
-using context = Cratis.Chronicle.Integration.Orleans.InProcess.for_Reactors.when_handling_event.and_it_fails_first_time.but_not_third_time.context;
+using Humanizer;
+using context = Cratis.Chronicle.Integration.Orleans.InProcess.for_Reactors.when_handling_event.and_it_fails.but_not_second_time.context;
 
-namespace Cratis.Chronicle.Integration.Orleans.InProcess.for_Reactors.when_handling_event.and_it_fails_first_time;
+namespace Cratis.Chronicle.Integration.Orleans.InProcess.for_Reactors.when_handling_event.and_it_fails;
 
 [Collection(GlobalCollection.Name)]
-public class but_not_third_time(context context) : Given<context>(context)
+public class but_not_second_time(context context) : Given<context>(context)
 {
-    public class context(GlobalFixture globalFixture) : given.a_reactor_observing_an_event_that_can_fail(globalFixture, 3)
+    public class context(GlobalFixture globalFixture) : given.a_reactor_observing_an_event_that_can_fail(globalFixture, 2)
     {
         public IEnumerable<FailedPartition> FailedPartitionsBeforeRetry;
         public IEnumerable<FailedPartition> FailedPartitionsAfterRetry;
@@ -28,27 +29,21 @@ public class but_not_third_time(context context) : Given<context>(context)
 
         async Task Because()
         {
+            var waitTime = 5.Seconds();
             await ReactorObserver.WaitTillActive();
             Observers[0].ShouldFail = true;
-            Observers[1].ShouldFail = true;
-            Observers[2].ShouldFail = false;
+            Observers[1].ShouldFail = false;
             await EventStore.EventLog.Append(EventSourceId, Event);
 
             // Wait for the first event to have been handled
-            await Tcs[0].Task.WaitAsync(TimeSpan.FromSeconds(60));
+            await Tcs[0].Task.WaitAsync(waitTime);
 
-            FailedPartitionsBeforeRetry = await EventStore.WaitForThereToBeFailedPartitions(ObserverId, TimeSpan.FromSeconds(60));
-            Jobs = await EventStore.WaitForThereToBeJobs(TimeSpan.FromSeconds(60));
+            FailedPartitionsBeforeRetry = await EventStore.WaitForThereToBeFailedPartitions(ObserverId);
+            Jobs = await EventStore.WaitForThereToBeJobs();
 
             // Wait for the second event to have been handled
-            await Tcs[1].Task.WaitAsync(TimeSpan.FromSeconds(60));
-
-            FailedPartitionsBeforeRetry = await EventStore.WaitForThereToBeFailedPartitions(ObserverId, TimeSpan.FromSeconds(60));
-            Jobs = await EventStore.WaitForThereToBeJobs(TimeSpan.FromSeconds(60));
-
-            // Wait for the third event to have been handled
-            await Tcs[2].Task.WaitAsync(TimeSpan.FromSeconds(60));
-            await EventStore.WaitForThereToBeNoJobs(TimeSpan.FromSeconds(60));
+            await Tcs[1].Task.WaitAsync(waitTime);
+            await EventStore.WaitForThereToBeNoJobs();
 
             FailedPartitionsAfterRetry = await GetFailedPartitions();
         }
